ADDED A FIREWALL
This acts as a firewall, also providing proxy functionalities such as
masking the client's IP address and filtering content, between an
incoming connection (external to the machine) and the local server.
The firewall listens on all network interfaces (0.0.0.0) at port 5100 and,
if the IP Address of the incoming connection is among those allowed,
then it forwards the packet to the local server.